Added file watching process restart option.
To use it, specify --watch_file=<filename> which means that the process that is being watched
must write or at least update the modified time on that file or it will be
restarted, presumed being frozen or hung.

Specify the watch period with --watch_file_seconds=<seconds> and a startup
delay with --watch_file_delay=<seconds>.

Using terrible command line parsing, so the command line format is
inflexible.

An example would be

forever.exe --watch_file=alive.txt --watch_file_seconds=20
--watch_file_delay=100 process_that_must_live.exe "--living_process_arg
blah --another"